% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDStoyCK-package.R
\docType{package}
\name{HDStoyCK-package}
\alias{HDStoyCK}
\alias{HDStoyCK-package}
\title{HDStoyCK: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristK/HDStoyCK}
  \item Report bugs at \url{https://github.com/ChristK/HDStoyCK/issues}
}

}
\author{
\strong{Maintainer}: Chris Kypridemos \email{christodoulosk@gmail.com} (\href{https://orcid.org/07896}{ORCID})

}
\keyword{internal}
