% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSumsDT.R
\name{rowSumsDT}
\alias{rowSumsDT}
\title{rowSumsDT}
\usage{
rowSumsDT(dtb, na.rm = FALSE)
}
\arguments{
\item{dtb}{A \code{data.table}.}

\item{na.rm}{If \code{TRUE} removes NAs from rows before the sum. Else returns NA for rows that contain NAs.}
}
\value{
A numeric vector if \code{dtb} contains numeric columns. Else \code{NULL}.
}
\description{
Form row sums for numeric columns in a data.table
}
\examples{
library(data.table)
library(HDStoyCK)
x <- matrix(fdeaths, ncol = 12, byrow = TRUE)
x <- data.table(x)
rowSumsDT(x)
}
