% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDStoyCK-package.R
\docType{package}
\name{HDStoyCK-package}
\alias{HDStoyCK}
\alias{HDStoyCK-package}
\title{HDStoyCK: Sums Rows In A data.table (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChristK/HDStoyCK}
  \item Report bugs at \url{https://github.com/ChristK/HDStoyCK/issues}
}

}
\author{
\strong{Maintainer}: Chris Kypridemos \email{first.last@example.com}

}
\keyword{internal}
